pyproject.toml
[tool.poetry]
name = "subpython"
version = "0.1.0"
description = "Use python process as module"
authors = ["Dalteo Craft <dalteo@dalteocraft.com>"]

[tool.poetry.dependencies]
python = ">=3.6"
pyzmq = "^20.0.0"
tornado = "^6.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
subpython = "subpython.__main__:main"
